+++ slang/lib/builder/builder/generate_config_builder.dart
@@ -1,4 +1,5 @@
 import 'package:slang/builder/builder/build_model_config_builder.dart';
+import 'package:slang/builder/model/context_type.dart';
 import 'package:slang/builder/model/raw_config.dart';
 import 'package:slang/builder/model/generate_config.dart';
 import 'package:slang/builder/model/interface.dart';
@@ -7,6 +8,7 @@ class GenerateConfigBuilder {
   static GenerateConfig build({
     required String baseName,
     required RawConfig config,
+    required List<ContextType> contexts,
     required List<Interface> interfaces,
   }) {
     return GenerateConfig(
@@ -23,7 +25,13 @@ class GenerateConfigBuilder {
       renderFlatMap: config.renderFlatMap,
       translationOverrides: config.translationOverrides,
       renderTimestamp: config.renderTimestamp,
-      contexts: config.contexts,
+      contexts: contexts.map((c) {
+        return PopulatedContextType(
+          enumName: c.enumName,
+          enumValues: c.enumValues!,
+          generateEnum: c.generateEnum,
+        );
+      }).toList(),
       interface: interfaces,
       obfuscation: config.obfuscation,
       imports: config.imports,