+++ slang/lib/builder/builder/raw_config_builder.dart
@@ -121,7 +121,7 @@ extension on Map<String, dynamic> {
 
       return ContextType(
         enumName: enumName,
-        enumValues: config['enum'].cast<String>(),
+        enumValues: config['enum']?.cast<String>(),
         paths: config['paths']?.cast<String>() ?? ContextType.defaultPaths,
         defaultParameter:
             config['default_parameter'] ?? ContextType.DEFAULT_PARAMETER,