+++ slang/lib/builder/model/i18n_data.dart
@@ -1,4 +1,4 @@
-import 'package:slang/builder/builder/translation_model_builder.dart';
+import 'package:slang/builder/model/context_type.dart';
 import 'package:slang/builder/model/i18n_locale.dart';
 import 'package:slang/builder/model/interface.dart';
 import 'package:slang/builder/model/node.dart';
@@ -10,12 +10,14 @@ class I18nData {
   final bool base; // whether or not this is the base locale
   final I18nLocale locale; // the locale (the part after the underscore)
   final ObjectNode root; // the actual strings
-  final List<Interface> interfaces;
+  final List<ContextType> contexts; // detected context types
+  final List<Interface> interfaces; // detected interfaces
 
   I18nData({
     required this.base,
     required this.locale,
     required this.root,
+    required this.contexts,
     required this.interfaces,
   });
 
@@ -29,18 +31,4 @@ class I18nData {
       return -1; // move base to the left
     }
   };
-}
-
-extension I18nBuilderExt on BuildModelResult {
-  I18nData toI18nData({
-    required bool base,
-    required I18nLocale locale,
-  }) {
-    return I18nData(
-      base: base,
-      locale: locale,
-      root: root,
-      interfaces: interfaces,
-    );
-  }
 }